+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.4.0] 2024-09-25
+
+### Added
+
+* Added system slash command to manually invoke tools.
+
 ## [5.3.0] 2024-09-24
 
 ### Added
 
-* New ChatAgent parameter default_max_response_tokens
+* New ChatAgent parameter default_max_response_tokens.
 
 ## [5.2.2] 2024-09-19